Rename action-trust.js to action-constants.js
