+++ extensions/amp-date-picker/0.1/amp-date-picker.js
@@ -15,7 +15,7 @@
  */
 
 import '../../../third_party/react-dates/bundle';
-import {ActionTrust} from '../../../src/action-trust';
+import {ActionTrust} from '../../../src/action-constants';
 import {AmpEvents} from '../../../src/amp-events';
 import {CSS} from '../../../build/amp-date-picker-0.1.css';
 import {DEFAULT_FORMAT, DEFAULT_LOCALE, FORMAT_STRINGS} from './constants';
