+++ extensions/amp-gwd-animation/0.1/amp-gwd-animation.js
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import {ActionTrust} from '../../../src/action-trust';
+import {ActionTrust} from '../../../src/action-constants';
 import {
   AmpGwdRuntimeService,
   GWD_SERVICE_NAME,
