+++ extensions/amp-lightbox/0.1/amp-lightbox.js
@@ -15,7 +15,7 @@
  */
 
 import * as st from '../../../src/style';
-import {ActionTrust} from '../../../src/action-trust';
+import {ActionTrust} from '../../../src/action-constants';
 import {AmpEvents} from '../../../src/amp-events';
 import {CSS} from '../../../build/amp-lightbox-0.1.css';
 import {Gestures} from '../../../src/gesture';
