+++ extensions/amp-position-observer/0.1/amp-position-observer.js
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {ActionTrust} from '../../../src/action-trust';
+import {ActionTrust} from '../../../src/action-constants';
 import {
   PositionObserverFidelity,
 } from '../../../src/service/position-observer/position-observer-worker';
