+++ src/service/video-manager-impl.js
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-import {ActionTrust} from '../action-trust';
+import {ActionTrust} from '../action-constants';
 import {
   EMPTY_METADATA,
   parseFavicon,
